Minor refactory of file tree
