+++ src/screens/index.js
@@ -1,7 +1,5 @@
-export { default as LoginScreen } from './LoginScreen/LoginScreen';
+export { default as LoginScreen } from "./LoginScreen/LoginScreen";
 
-export { default as HomeScreen } from './HomeScreen/HomeScreen';
+export { default as HomeScreen } from "./HomeScreen/HomeScreen";
 
-export { default as RegistrationScreen } from './RegistrationScreen/RegistrationScreen';
-
-export {default as MapScreen} from './MapScreen/MapScreen'
+export { default as RegistrationScreen } from "./RegistrationScreen/RegistrationScreen";
